move github req handler into untils
